refactor: organize per-product
now that we understand the api better,
I'm breaking the components and hooks into
per-product folders so that they're easier to read/test/maintain
