+++ reactfire/database/database.test.tsx
@@ -1,0 +1,11 @@
+import 'jest-dom/extend-expect';
+
+describe('Realtime Database (RTDB)', () => {
+  describe('useDatabaseObject', () => {
+    test.todo("returns the same value as ref.on('value')");
+  });
+
+  describe('useDatabaseList', () => {
+    test.todo("returns the same value as ref.on('value')");
+  });
+});
