+++ reactfire/firestore/firestore.test.tsx
@@ -1,0 +1,13 @@
+import { renderHook, act } from '@testing-library/react-hooks';
+import * as React from 'react';
+import 'jest-dom/extend-expect';
+
+describe('Firestore', () => {
+  describe('useFirestoreDoc', () => {
+    test.todo('returns the same value as ref.onSnapshot()');
+  });
+
+  describe('useFirestoreCollection', () => {
+    test.todo('returns the same value as ref.onSnapshot()');
+  });
+});
