+++ reactfire/storage/storage.test.tsx
@@ -1,0 +1,11 @@
+import 'jest-dom/extend-expect';
+
+describe('Storage', () => {
+  describe('useStorageTask', () => {
+    test.todo('returns the same value as uploadTask');
+  });
+
+  describe('useStorageDownloadURL', () => {
+    test.todo('returns the same value as  getDownloadURL');
+  });
+});
